Move code generation behind a trait
