+++ crates/mockiato-codegen/src/code_generator/bound_lifetimes.rs
