+++ crates/mockiato-codegen/src/code_generator/lifetime_rewriter.rs
