+++ crates/mockiato-codegen/src/controller_impl.rs
@@ -1,5 +1,5 @@
+use crate::code_generator::{CodeGenerator, GenerateMockOptions};
 use crate::diagnostic::DiagnosticBuilder;
-use crate::generate::{generate_mock, GenerateMockOptions};
 use crate::parse::mockable_attr_parser::{MockableAttrParser, RemoteTraitPath};
 use crate::parse::trait_decl::TraitDecl;
 use crate::result::{Error, Result};
@@ -12,12 +12,17 @@ use syn::{AttributeArgs, Item, ItemTrait};
 #[derive(Debug)]
 pub(crate) struct ControllerImpl {
     mockable_attr_parser: Box<dyn MockableAttrParser>,
+    code_generator: Box<dyn CodeGenerator>,
 }
 
 impl ControllerImpl {
-    pub(crate) fn new(mockable_attr_parser: Box<dyn MockableAttrParser>) -> Self {
+    pub(crate) fn new(
+        mockable_attr_parser: Box<dyn MockableAttrParser>,
+        code_generator: Box<dyn CodeGenerator>,
+    ) -> Self {
         Self {
             mockable_attr_parser,
+            code_generator,
         }
     }
 }
@@ -33,17 +38,15 @@ impl Controller for ControllerImpl {
             None => Some(item_trait),
         };
 
-        let generated_mock = generate_mock(
-            &trait_decl,
-            GenerateMockOptions {
-                custom_struct_ident: mockable_attr.name,
-                force_static_lifetimes: mockable_attr.force_static_lifetimes,
-                custom_trait_path: match mockable_attr.remote_trait_path {
-                    Some(RemoteTraitPath::Path(path)) => Some(path),
-                    _ => None,
-                },
+        let options = GenerateMockOptions {
+            custom_struct_ident: mockable_attr.name,
+            force_static_lifetimes: mockable_attr.force_static_lifetimes,
+            custom_trait_path: match mockable_attr.remote_trait_path {
+                Some(RemoteTraitPath::Path(path)) => Some(path),
+                _ => None,
             },
-        );
+        };
+        let generated_mock = self.code_generator.generate(&trait_decl, options);
 
         Ok(quote! {
             #emit_item_trait
